pyproject.toml
[project]
name = "hybrid-technique-for-image-encryption"
version = "0.1.0"
description = "This project implements a hybrid image encryption technique by combining asymmetric and symmetric encryption methods. It utilizes the ECC for asymmetric encryption and ChaCha20 stream cipher for symmetric encryption to achieve robust and secure image protection."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"ecpy>=1.2.5",
	"numpy>=2.3.2",
	"pillow>=11.3.0",
	"pycryptodome>=3.23.0",
	"python-dotenv>=1.1.1",
]
